В описании...
Вынес notification в отдельную папку Services. Починил от вылета при создании таски без уведомления. Поправил нейминг
